+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix robustness to very low desired tolerance in `wp.fem.utils.symmetric_eigenvalues_qr`
 - Fix invalid code generation error messages when nesting dynamic and static for-loops
 - Fix caching of kernels with static expressions
+- Fix `ModelBuilder.add_builder(builder)` to correctly update `articulation_start` and thereby `articulation_count` when `builder` contains more than one articulation
 
 ## [1.4.0] - 2024-10-01